% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{compute_omega}
\alias{compute_omega}
\title{Deprecated: use \code{\link{compute_indices}} instead}
\usage{
compute_omega(results)
}
\arguments{
\item{results}{An \code{esem_comparison_pipeline} object.}
}
\value{
A \code{reliability_indices} object.
}
\description{
Deprecated: use \code{\link{compute_indices}} instead
}
