% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_syntax}
\alias{get_syntax}
\title{Extract the Generated lavaan Syntax}
\usage{
get_syntax(x, cat = TRUE)
}
\arguments{
\item{x}{An \code{esem_fit} object.}

\item{cat}{Logical. If \code{TRUE} (default), also print to console.}
}
\value{
The lavaan model string (invisibly).
}
\description{
Extract the Generated lavaan Syntax
}
