% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print.esem_spec}
\alias{print.esem_spec}
\title{Print an esem_spec Object}
\usage{
\method{print}{esem_spec}(x, ...)
}
\arguments{
\item{x}{An \code{esem_spec} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}; called for the side effect of printing
the model specification.
}
\description{
Print an esem_spec Object
}
